pyproject.toml
[tool.poetry]
name = "upwork-job-scraper"
version = "1.0.0"
description = "Production-grade Upwork job scraper with AI cover letter generation"
authors = ["Developer <developer@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
crawl4ai = ">=0.4.24"
pandas = "^2.1.0"
openpyxl = "^3.1.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"
black = "^23.0.0"
mypy = "^1.5.0"
types-pyyaml = "^6.0.0"

[tool.poetry.scripts]
scrape = "src.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
